ONT-3068- jitter and packet loss results sent to server, tested

+++ core/src/main/java/at/specure/data/entity/TestRecord.kt
@@ -254,5 +254,5 @@ data class TestRecord(
     /**
      * packet loss in % 99.3
      */
-    var packetLossPercents: Float? = null
+    var packetLossPercents: Double? = null
 )
